crt0_app：把setup移动到crt0里面
