+++ vk.application/console/src/Console.cpp
@@ -255,19 +255,6 @@ const char* Console::GetPath()
 }
 
 
-/// @brief Kernel
-Kernel* kernel;
-
-
-/// @brief setup
-/// @param kernel 
-/// @return 
-extern "C" void setup(void* knl)
-{
-	kernel = (Kernel*)knl;
-}
-
-
 /// @brief main
 int main(int argc, char* argv[])
 {
