+++ vk.application/launch/src/Launch.cpp
@@ -83,19 +83,6 @@ void Launch::Execute()
 }
 
 
-/// @brief Kernel
-Kernel* kernel;
-
-
-/// @brief setup
-/// @param kernel 
-/// @return 
-extern "C" void setup(void* knl)
-{
-	kernel = (Kernel*)knl;
-}
-
-
 /// @brief main
 int main(int argc, char* argv[])
 {
